splitting up quiz into different components
